refactor code for highter unit test coverage
