+++ prog-2-teamAF/src/main/java/com/teamAF/app/Controller/HomeController.java
@@ -43,10 +43,6 @@ public class HomeController implements Initializable {
     public HomeController() {
     }
 
-    public HomeController(MovieService movieService) {
-        this.movieService = movieService;
-    }
-
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
         if (movieService == null) {
